chore, obey icc warnings

+++ src/fix.c
@@ -261,7 +261,9 @@ fixc_render_fld(
 		break;
 	}
 	case FIXC_TYP_UCHAR:
-		res += snprintf(buf + res, bsz - res, "%03" PRIu8, fld.u8);
+		res += snprintf(
+			buf + res, bsz - res,
+			"%03u", (unsigned int)fld.u8);
 		break;
 	case FIXC_TYP_CHAR:
 		buf[res++] = fld.c;
